Should have done this in multiple commits
Leaving this here now however this is still WIP.
Changed the layout of the brackens_renderer to have folders for each area of the library: pipelines, textures, models. and have changed the import paths in the brackens_engine library to match this.
Started adding the model renderer however needed to create a new pipeline to do so as we're not using simple instance pipeline. Added the wgsl code also (still WIP).
As such, am currently creating a new RawPipeline struct. So far, i believe that this is mostly redundant outside of saving a little bit of time when creating the pipeline.