Created RawPipeline
I think this is mostly finished. There may be a few other things to add later but this is good enough for now. Going to change instance pipeline to be a wrapper around this as much of the code is the same. If i've designed this correctly, it should be fairly straight forward.