Updated naming convension
View and ViewMut when in systems should be prefixed with v_ or vm_. Not 100% on this change but it keeps everything more inline. Uniques are left unchanged as they're fine as they are.
Also removed UV, UVM, CV and CVM as massively uncoordinated in their application and code looks generally nicer without them.